feat: mock jwt util
