+++ rendezvous-rest/src/main/kotlin/com/zoin/rendezvous/config/JwtUtilConfig.kt
@@ -1,0 +1,4 @@
+package com.zoin.rendezvous.config
+
+class JwtUtilConfig {
+}
